Edit plate reader embed middleware

+++ app/src/discord/events/definitions/numbersGamePlateReader.ts
@@ -16,6 +16,10 @@ export default defineEvent({
       return;
     }
 
+    if (await handleEmbedModify(message)) {
+      return;
+    }
+
     const attachment = message.attachments.first();
     if (!attachment) return;
     if (!attachment.contentType?.startsWith("image/")) return;
@@ -78,6 +82,61 @@ const handleEmbedRemove = async (message: Message): Promise<boolean> => {
   await Promise.all([
     refMessage.edit({
       embeds: refMessage.embeds.filter((_el, index) => index !== embedNr - 1),
+    }),
+    message.delete(),
+  ]);
+
+  return true;
+};
+
+const handleEmbedModify = async (message: Message): Promise<boolean> => {
+  const { member, content, reference, channel } = message;
+
+  if (
+    !member?.permissions.has(PermissionFlagsBits.Administrator) ||
+    !reference?.messageId
+  ) {
+    return false;
+  }
+
+  const embedNrRegExp = /^edit embed (\d+) (.+)$/;
+  const matches = embedNrRegExp.exec(content);
+
+  if (!matches) {
+    return false;
+  }
+
+  const embedNr = parseInt(matches[1] ?? "");
+  if (!Number.isFinite(embedNr)) {
+    return false;
+  }
+
+  const newContent = matches[2] ?? "";
+  if (!newContent) {
+    return false;
+  }
+
+  const refMessage = channel.messages.cache.get(reference.messageId);
+  if (!refMessage || refMessage.author.id !== message.client.user.id) {
+    return false;
+  }
+
+  if (typeof refMessage.embeds[embedNr - 1] === "undefined") {
+    return false;
+  }
+
+  await Promise.all([
+    refMessage.edit({
+      embeds: refMessage.embeds.map((el, index) => {
+        if (index !== embedNr - 1) {
+          return el;
+        }
+
+        return {
+          ...el,
+          description: newContent,
+        };
+      }),
     }),
     message.delete(),
   ]);
